requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.6.0
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.3